Refactor InstanceCreationPage to use GeneralInfoForm in other places
